requirements.txt
accelerate==0.27.2
datasets==2.18.0
graphviz==0.20.1
huggingface-hub==0.21.3
matplotlib==3.8.2
numpy==1.24.4
pandas==2.2.1
pytest==7.4.0
pyyaml==6.0.1
rich==13.5.2
scikit-learn==1.3.0
seaborn==0.12.2
torchinfo==1.8.0
torchview==0.2.6
tqdm==4.66.1
transformers==4.38.1
hydra-core==1.3.2
wandb==0.16.3
